webpack is installed, some componenets are separated
