+++ homework7/db.json
@@ -2,38 +2,17 @@
   "tasks": [
     {
       "task": "tas1",
-      "categories": "work",
-      "date": "2023-04-16",
-      "completed": true,
+      "categories": "home",
+      "date": "2023-03-28",
+      "completed": false,
       "id": 1
     },
     {
-      "task": "davaleab1",
+      "task": "tas2",
       "categories": "work",
-      "date": "2023-04-11",
-      "completed": true,
+      "date": "2023-04-12",
+      "completed": false,
       "id": 2
-    },
-    {
-      "task": "gav",
-      "categories": "other",
-      "date": "2023-04-03",
-      "completed": true,
-      "id": 3
-    },
-    {
-      "task": "qASD",
-      "categories": "home",
-      "date": "2023-03-28",
-      "completed": true,
-      "id": 4
-    },
-    {
-      "task": "tas",
-      "categories": "work",
-      "date": "2023-04-11",
-      "completed": true,
-      "id": 5
     }
   ]
 }
