+++ homework7/webpack.config.js
@@ -1,0 +1,34 @@
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const path = require("path");
+
+module.exports = {
+  mode: process.env.NODE_ENV || "development",
+  output: {
+    clean: true,
+    filename: "bundle.js",
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: "index.html",
+    }),
+  ],
+  resolve: {
+    extensions: [".js", ".jsx"],
+    modules: [path.resolve(__dirname, "src"), "node_modules"],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        use: {
+          loader: "babel-loader",
+        },
+      },
+      {
+        test: /\.css$/i,
+        use: ["style-loader", "css-loader"],
+      },
+    ],
+  },
+};
